+++ test.sh
@@ -33,6 +33,8 @@ rna_pdb_toolsx.py --no-hr --clean input/1osw.pdb > output/1osw_NMR_1.pdb
 rna_pdb_toolsx.py --no-hr --get-rnapuzzle-ready input/1xjr_onlyGTP.pdb > input/1xjr_onlyGTP_X.pdb
 rna_pdb_toolsx.py --no-hr --get-rnapuzzle-ready input/1xjr_onlyGTP.pdb > output/1xjr_onlyGTP_rnapuzzle_ready.pdb
 rna_pdb_toolsx.py --no-hr --get-rnapuzzle-ready input/1_das_1_rpr_fixed.pdb > output/1_das_1_rpr_fixed.pdb
+rna_pdb_toolsx.py --no-hr --inspect input/1_das_1_rpr_fixed.pdb > output/1_das_1_rpr_inspect.txt
+
 rna_pdb_toolsx.py --no-hr --get-rnapuzzle-ready input/4GXY_3firstNt.pdb > output/4GXY_3firstNt.pdb
 rna_pdb_toolsx.py --no-hr --get-rnapuzzle-ready input/gtp.pdb  > output/gtp.pdb
 rna_pdb_toolsx.py --no-hr --get-rnapuzzle input/377D.pdb > output/377D.txt # should finish with error