+++ CHANGELOG.md
@@ -6,7 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 # Master
 
-- rna_pdb_toolsx.py: add --backbone-only
+- rna_pdb_toolsx.py:
+     - add --inspect
+	 - add --backbone-only
 - rna_calc_rmsd_biopython.py: 
      - introduce way to save models after rmsd calc
 	 - add --triple-mode,--column-name