+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - rna_pdb_toolsx.py:
      - add --inspect
 	 - add --backbone-only
+- rna_tools_lib.py: 
+     - add reload()
 - rna_calc_rmsd_biopython.py: 
      - introduce way to save models after rmsd calc
 	 - add --triple-mode,--column-name