+++ CHANGELOG.md
@@ -6,16 +6,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 # Master
 
+rna_tools_lib.py (RNAStructure) - get_res_text()
+rna_pdb_toolsx.py: --set-chain; set chain for all ATOM lines and TER (quite brutal function)
+
 - rna_pdb_toolsx.py:
      - add --inspect
 	 - add --backbone-only
+
 - rna_tools_lib.py: 
      - RNAStructure:
 	   - add reload()
 	   - change write() now saves by default to self.fn
        - add get_remarks_text()
 	   - add get_res_text()
-
 	   
 - rna_calc_rmsd_biopython.py: 
      - introduce way to save models after rmsd calc