+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 rna_tools_lib.py (RNAStructure) - get_res_text()
 rna_pdb_toolsx.py: --set-chain, set chain for all ATOM lines and TER (quite brutal function)
+rna_calc_rmsd_biopython.py: --triple-mode works! 201118
 
 - rna_pdb_toolsx.py:
      - add --inspect
@@ -23,6 +24,8 @@ rna_pdb_toolsx.py: --set-chain, set chain for all ATOM lines and TER (quite brut
 - rna_calc_rmsd_biopython.py: 
      - introduce way to save models after rmsd calc
 	 - add --triple-mode,--column-name
+
+
 
 # 3.5.4 200710